remove old tools page
