--- frontend/tools/components/__tests__/toolbay_header_test.tsx
@@ -1,10 +1,0 @@
-import * as React from "react";
-import { ToolBayHeader } from "../toolbay_header";
-import { mount } from "enzyme";
-
-describe("<ToolBayHeader />", () => {
-  it("renders", () => {
-    const header = mount(<ToolBayHeader />);
-    expect(header.text()).toEqual("SlotXYZTool or Seed Container");
-  });
-});
