--- frontend/tools/components/index.ts
@@ -1,4 +1,0 @@
-export * from "./toolbay_form";
-export * from "./toolbay_list";
-export * from "./tool_list";
-export * from "./tool_form";
